got markers to render from database

+++ db/seed.sql
@@ -23,23 +23,7 @@ CREATE TABLE emails(
 
 CREATE TABLE positions(
     position_id SERIAL PRIMARY KEY,
-    lat INTEGER,
-    lng INTEGER,
+    lat DECIMAL,
+    lng DECIMAL,
     username VARCHAR(100)
 );
-
--- DROP TABLE IF EXISTS user_cart_junction;
--- DROP TABLE IF EXISTS carts;
-
--- CREATE TABLE carts(
---     cart_id SERIAL PRIMARY KEY,
---     user_id INT REFERENCES users(user_id),
---     active BOOLEAN
--- );
-
--- CREATE TABLE user_cart_junction(
---     user_cart_id SERIAL PRIMARY KEY,
---     cart_id INT REFERENCES carts(cart_id),
---     product_id INT REFERENCES products(product_id),
---     quantity INT
--- );
